feat: useAiFeedbackWatcher
