+++ src/components/reflection-detail/article/ReflectionArticle.tsx
@@ -1,10 +1,9 @@
-import { useEffect, useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
 import { Box, Typography } from "@mui/material";
 import { label } from "../../post-form/popup/select-tag/button/TagButton";
 import { StyledMarkdown } from "./mark-down";
-import supabase from "@/src/lib/supabase";
+import { useAiFeedbackWatcher } from "@/src/hooks/reflection/useAiFeedbackWatcher";
 import { formatDate } from "@/src/utils/date-helper";
 import { theme } from "@/src/utils/theme";
 
@@ -36,10 +35,6 @@ const h1 = {
   outline: "none"
 };
 
-type AiFeedback = {
-  aiFeedback: string;
-};
-
 export const ReflectionArticle: React.FC<ReflectionArticleProps> = ({
   username,
   userImage,
@@ -50,56 +45,7 @@ export const ReflectionArticle: React.FC<ReflectionArticleProps> = ({
   activeTags,
   reflectionCUID
 }) => {
-  const [aiFeedback, setAiFeedback] = useState<string>("まじで出ろ");
-
-  // 初期データ取得
-  const fetchComments = async () => {
-    const { data, error } = await supabase
-      .from("Reflection")
-      .select("aiFeedback")
-      .eq("reflectionCUID", reflectionCUID)
-      .single();
-
-    if (error) {
-      console.error("Error fetching comments", error);
-    } else {
-      setAiFeedback(data.aiFeedback || "まじで出ろ");
-    }
-  };
-
-  // リアルタイム更新購読
-  const subscribeToRealtimeUpdates = () => {
-    const channel = supabase
-      .channel("realtime:ai")
-      .on<AiFeedback>(
-        "postgres_changes",
-        {
-          event: "UPDATE",
-          schema: "public",
-          table: "Reflection",
-          filter: `reflectionCUID=eq.${reflectionCUID}`
-        },
-        (payload) => {
-          if (payload.new && payload.new.aiFeedback) {
-            setAiFeedback(payload.new.aiFeedback);
-          }
-        }
-      )
-      .subscribe();
-
-    return channel;
-  };
-
-  useEffect(() => {
-    fetchComments();
-    const channel = subscribeToRealtimeUpdates();
-
-    return () => {
-      supabase.removeChannel(channel);
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
+  const aiFeedback = useAiFeedbackWatcher(reflectionCUID);
   const handleSendToSQS = async () => {
     await onSendToSQS();
   };
